move out of findeligiblepayments package
