+++ ShopperInsights/src/main/java/com/braintreepayments/api/FindEligiblePaymentsApi.kt
@@ -1,6 +1,6 @@
-package com.braintreepayments.api.findeligiblepayments
+package com.braintreepayments.api
 
-import com.braintreepayments.api.findeligiblepayments.FindEligiblePaymentsApiRequest.Companion.toJson
+import com.braintreepayments.api.FindEligiblePaymentsApiRequest.Companion.toJson
 
 internal class FindEligiblePaymentsApi {
     fun execute(request: FindEligiblePaymentsApiRequest): FindEligiblePaymentsApiResult {
