+++ ShopperInsights/src/main/java/com/braintreepayments/api/ShopperInsightsApi.kt
@@ -1,9 +1,5 @@
 package com.braintreepayments.api
 
-import com.braintreepayments.api.findeligiblepayments.FindEligiblePaymentsApiResult
-import com.braintreepayments.api.findeligiblepayments.FindEligiblePaymentsApi
-import com.braintreepayments.api.findeligiblepayments.FindEligiblePaymentsApiRequest
-
 /**
  * [ShopperInsightsApi] is a wrapper class for all api related classes that
  * [ShopperInsightsClient] depends on.
